Semantic annotation tests separation
