--- testSrc/com/intellij/plugins/haxe/ide/HaxeSemanticAnnotatorTest.java
@@ -1,255 +1,0 @@
-/*
- * Copyright 2000-2013 JetBrains s.r.o.
- * Copyright 2014-2015 AS3Boyan
- * Copyright 2014-2014 Elias Ku
- * Copyright 2018 Ilya Malanin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.intellij.plugins.haxe.ide;
-
-import com.intellij.codeInsight.intention.IntentionAction;
-import com.intellij.lang.LanguageAnnotators;
-import com.intellij.openapi.fileEditor.FileDocumentManager;
-import com.intellij.plugins.haxe.HaxeCodeInsightFixtureTestCase;
-import com.intellij.plugins.haxe.HaxeLanguage;
-import com.intellij.plugins.haxe.ide.annotator.HaxeTypeAnnotator;
-import com.intellij.util.ArrayUtil;
-import org.apache.commons.lang.StringUtils;
-
-import java.util.Arrays;
-
-public class HaxeSemanticAnnotatorTest extends HaxeCodeInsightFixtureTestCase {
-  @Override
-  protected String getBasePath() {
-    return "/annotation.semantic/";
-  }
-
-  private void doTestNoFix(boolean checkWarnings, boolean checkInfos, boolean checkWeakWarnings, String... additionalFiles) throws Exception {
-    myFixture.configureByFiles(ArrayUtil.mergeArrays(new String[]{getTestName(false) + ".hx"}, additionalFiles));
-    final HaxeTypeAnnotator annotator = new HaxeTypeAnnotator();
-    LanguageAnnotators.INSTANCE.addExplicitExtension(HaxeLanguage.INSTANCE, annotator);
-    myFixture.enableInspections(getAnnotatorBasedInspection());
-    myFixture.testHighlighting(true, false, true);
-  }
-
-  private void doTestNoFixWithWarnings(String... additionalFiles) throws Exception {
-    doTestNoFix(true, false, false, additionalFiles);
-  }
-
-  private void doTestNoFixWithoutWarnings(String... additionalFiles) throws Exception {
-    doTestNoFix(false, false, false, additionalFiles);
-  }
-
-  private void doTest(String... filters) throws Exception {
-    doTestNoFixWithoutWarnings();
-    for (final IntentionAction action : myFixture.getAvailableIntentions()) {
-      if (Arrays.asList(filters).contains(action.getText())) {
-        System.out.println("Applying intent " + action.getText());
-        myFixture.launchAction(action);
-      } else {
-        System.out.println("Ignoring intent " + action.getText() + ", not matching " + StringUtils.join(filters, ","));
-      }
-    }
-    FileDocumentManager.getInstance().saveAllDocuments();
-    myFixture.checkResultByFile(getTestName(false) + "_expected.hx");
-  }
-
-  public void testFixPackage() throws Exception {
-    doTest("Set package name to ''");
-  }
-
-  public void testRemoveOverride() throws Exception {
-    doTest("Drop modifier 'override'");
-  }
-
-  public void testRemoveFinal() throws Exception {
-    doTest("Drop modifier '@:final' from 'Base.test'");
-  }
-
-  public void testChangeArgumentType() throws Exception {
-    doTest("Change parameter 'a' type from 'Int' to 'Bool'");
-  }
-
-  public void testRemoveArgumentInit() throws Exception {
-    doTest("Drop initializer");
-  }
-
-  public void testInterfaceMethodsShouldHaveTypeTags() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testOptionalWithInitWarning() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testNonConstantArgument() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testNonConstantArgumentAbstractEnum() throws Exception {
-    doTestNoFixWithWarnings("test/SampleAbstractEnum.hx", "std/StdTypes.hx");
-  }
-
-  public void testConstructorMustNotBeStatic() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testInitMagicMethodShouldBeStatic() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testRepeatedArgumentName() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testAbstractFromTo() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testAbstractAssignmentFromTo1() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testAbstractAssignmentFromTo2() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testAbstractAssignmentFromTo3() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testAbstractAssignmentFromTo4() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testAbstractAssignmentFromTo5() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testAbstractAssignmentFromTo6() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testNullFunction() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testOverrideVisibility() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testUcFirstClassName() throws Exception {
-    doTest("Rename to 'Test'");
-  }
-
-  public void testUcFirstClassName2() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testRepeatedFields() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testPropertiesSimpleCheck() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testPropertyAllowNonConstantInitialization() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testOverrideSignature() throws Exception {
-    doTest("Remove parameter");
-  }
-
-  public void testOverrideSignature2() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testOverrideSignature3() throws Exception {
-    doTest("Remove parameter");
-  }
-
-  public void testOverrideSignature4() throws Exception {
-    doTest("Remove parameter");
-  }
-
-  public void testImplementSignature() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testImplementExternInterface() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testSimpleAssignUnknownGeneric() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testExtendsAnonymousType() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testExtendsSelf() throws Exception {
-    doTestNoFixWithWarnings("test/Bar.hx", "test/IBar.hx", "test/TBar.hx");
-  }
-
-  public void testFieldInitializerCheck() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testVariableRedefinition() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testFinalKeyword() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testArrayAssignmentFromEmpty() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testArrayAssignmentBadFunctionType() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testArrayAssignmentWrongType() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testArrayAssignmentBadArrowFunction() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testArrayAssignmentWithAbstract() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testArrayAssignmentWithArrowFunctions() throws Exception {
-    doTestNoFixWithWarnings("std/StdTypes.hx", "std/Array.hx");
-  }
-
-  public void testNullTAssignment1() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testNullTAssignment2() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-
-  public void testStringLiteral() throws Exception {
-    doTestNoFixWithWarnings();
-  }
-}
