+++ testSrc/com/intellij/plugins/haxe/ide/semantic/HaxeBaseSemanticAnnotatorTestCase.java
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2000-2013 JetBrains s.r.o.
+ * Copyright 2014-2018 AS3Boyan
+ * Copyright 2014-2014 Elias Ku
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.intellij.plugins.haxe.ide.semantic;
+
+import com.intellij.codeInsight.intention.IntentionAction;
+import com.intellij.lang.LanguageAnnotators;
+import com.intellij.openapi.fileEditor.FileDocumentManager;
+import com.intellij.plugins.haxe.HaxeCodeInsightFixtureTestCase;
+import com.intellij.plugins.haxe.HaxeLanguage;
+import com.intellij.plugins.haxe.ide.annotator.HaxeTypeAnnotator;
+import com.intellij.util.ArrayUtil;
+import org.apache.commons.lang.StringUtils;
+
+import java.util.Arrays;
+
+abstract class HaxeBaseSemanticAnnotatorTestCase extends HaxeCodeInsightFixtureTestCase {
+  @Override
+  protected String getBasePath() {
+    return "/annotation.semantic/";
+  }
+
+  private void doTestNoFix(boolean checkWarnings, boolean checkInfos, boolean checkWeakWarnings, String... additionalFiles) throws Exception {
+    myFixture.configureByFiles(ArrayUtil.mergeArrays(new String[]{getTestName(false) + ".hx"}, additionalFiles));
+    final HaxeTypeAnnotator annotator = new HaxeTypeAnnotator();
+    LanguageAnnotators.INSTANCE.addExplicitExtension(HaxeLanguage.INSTANCE, annotator);
+    myFixture.enableInspections(getAnnotatorBasedInspection());
+    myFixture.testHighlighting(true, false, true);
+  }
+
+  protected void doTestNoFixWithWarnings(String... additionalFiles) throws Exception {
+    doTestNoFix(true, false, false, additionalFiles);
+  }
+
+  private void doTestNoFixWithoutWarnings(String... additionalFiles) throws Exception {
+    doTestNoFix(false, false, false, additionalFiles);
+  }
+
+  protected void doTest(String... filters) throws Exception {
+    doTestNoFixWithoutWarnings();
+    for (final IntentionAction action : myFixture.getAvailableIntentions()) {
+      if (Arrays.asList(filters).contains(action.getText())) {
+        System.out.println("Applying intent " + action.getText());
+        myFixture.launchAction(action);
+      } else {
+        System.out.println("Ignoring intent " + action.getText() + ", not matching " + StringUtils.join(filters, ","));
+      }
+    }
+    FileDocumentManager.getInstance().saveAllDocuments();
+    myFixture.checkResultByFile(getTestName(false) + "_expected.hx");
+  }
+}
